Adds 'en' as the default language when app bootstraps
* This fixes a bug where the translation library recieves an invalid valid for the user's
preferred language. This occurs when the user visits the homepage and a session is initiated,
but the user does not select a preferred language. Since a language has never been saved, the
translation library will try to load strings with an invalid language. This causes the user
to see the translation lookup keys, instead of the translated values